Make the internal fetch task available to our reference apps
Towards: #275
