+++ designcompose/build.gradle.kts
@@ -26,7 +26,6 @@ plugins {
     id("designcompose.conventions.publish.android")
     id("designcompose.conventions.roborazzi")
     id("designcompose.conventions.android-test-devices")
-    id("com.android.designcompose.internal")
 }
 
 @Suppress("UnstableApiUsage")
