--- plugins/gradle-plugin-internal/build.gradle.kts
@@ -1,38 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-plugins {
-    kotlin("jvm")
-    `java-gradle-plugin`
-    id("designcompose.conventions.base")
-}
-
-gradlePlugin {
-    plugins {
-        create("designcompose-gradle-plugin") {
-            id = "com.android.designcompose.internal"
-            displayName = "com.android.designcompose.internal.gradle.plugin"
-            implementationClass = "com.android.designcompose.gradle.internal.InternalGradlePlugin"
-        }
-    }
-}
-
-dependencies {
-    compileOnly(libs.android.gradlePlugin)
-    implementation(project(":gradle-plugin"))
-    testImplementation(libs.google.truth)
-    testImplementation(libs.junit.jupiter)
-}
