+++ reference-apps/cluster-demo/app/build.gradle.kts
@@ -94,4 +94,5 @@ dependencies {
     testImplementation(libs.roborazzi.junit)
     testImplementation(libs.androidx.test.espresso.core)
     testImplementation(libs.androidx.compose.ui.test.junit4)
+    testImplementation(testFixtures(project(":designcompose")))
 }
