+++ reference-apps/helloworld/helloworld-app/build.gradle.kts
@@ -92,4 +92,5 @@ dependencies {
     testImplementation(libs.roborazzi.junit)
     testImplementation(libs.androidx.test.espresso.core)
     testImplementation(libs.androidx.compose.ui.test.junit4)
+    testImplementation(testFixtures(project(":designcompose")))
 }
